Now generates elf files with x86-64 code
...Sort of. Both the code gen and the elf file generation are highly limited,
full of placeholder code and hardcoded constants, offsets etc. It however does
generate an object file that gcc is able to link.

Also, this commit message actually refers to the previous commit. Git rebase
broke as the user.name was mistyped as use.name and I got tons of fatal errors
as a result. This commit has fixed that as well, so any future commits should
be good.