Error reporter now takes span rather than separate line/column/length
Accepting a span makes code cleaner in many places, as line, column and
length was already read from a span in almost everywhere.

NodeInfo is also now imported as Span in many places, and referred as
span in variable names. Full rename needs a pass where local variables
and function parameters are renamed as well.