Move e2e tests to external files, create test harness
Instead of running e2e tests as regular #[test]s, tests are defined in
.yaml files and loaded and run by a testing harness. This allows
creating and updating test cases without recompiling the program. This
also allows easier non-optimized vs optimized test runs in the future,
as test definitions can be reused.

As a note, repetitive test cases are autogenerated and git ignored. The
generation scripts are currently not run automatically and must be done
manually instead.