Allow using && in if statement and related changes
Remove condition reversal in TAC, instead do this at bytecode level.
This allows using && in conditions, as there is no need to implement
reversion for &&. As a result, use JumpIfFalse in if condition as well.
This requires fixes in optimizer, as dead code elimination incorrectly
removed else branch as dead.